ck: tool poc
Gross implementation of tooling
to save and list shell command lines.

The reason for this tool is my
need to save and maintain a series of fresh
commands each time I ramp up a new project.

The commands are organized through topics.
This helps to maintain and consult the list.